pyproject.toml
[project]
name = "tazmic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "crawl4ai>=0.6.3",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "markdown>=3.8",
    "python-dotenv>=1.1.0",
    "unstructured>=0.17.2",
]
